fix: initialize statistics log file for updates
The log file was not instantiated with a default log entry, which
resulted in breaking the code. The script was unable to grab lines from
an empty file (statistics.txt) and thus failed to update the statistics
dict. The statistics.txt has been initialized fittingly with a datetime
of the start of the project.